refactor: rename todoFunctionality
